# docs/swagger.json
{ "openapi": "3.0.0",
    "info":{
        "title":"Speckio Web App",
        "description":"This is an express API for the Speckio web app. Made by NHMG Development.",
        "version":"1.0.0"
    },
    "servers":[
        {
            "url": "https://localhost"
        }
    ],
    "components": {
      "securitySchemes": {
        "ApiKeyAuth": {
          "type": "apiKey",
          "in": "header",
          "name": "authorization"
        }
      }
    },
    "security": [
      { "ApiKeyAuth": [] }
    ],
    "paths":{
...
      "/result/disc": {
        "get": {
          "tags": [
            "Result"
          ],
          "description": "Return an disc results for a given user.",
          "parameters": [
            {
              "name": "userID",
              "in": "query",
              "description": "ID of user to search for",
              "required": true,
              "schema": {
                "type": "number",
                "example": 1
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "results": {
                        "type": "array",
                        "example": [
                          {
                            "userID": 1,
                            "typeName": "Dominance C",
                            "behaviour": "Dominance C Behavioural Insights",
                            "motivators": "Dominance C Motivators",
                            "strengths": "Dominance C Strengths",
                            "weaknesses": "Dominance C Weaknesses",
                            "reaction": "Dominance C Reaction Modes",
                            "communication": "Dominance C Communication"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "User results not found"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/result/team/eqi": {
        "get": {
          "tags": [
            "Result"
          ],
          "description": "Return an EQi results for a given team",
          "parameters": [
            {
              "name": "teamID",
              "in": "query",
              "description": "ID of team to search for",
              "required": false,
              "schema": {
                "type": "number",
                "example": 1
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "results": {
                        "type": "array",
                        "example": [
                          {
                            "userID": 1,
                            "categoryName": "Self-Regard",
                            "score": 115,
                            "descName": "High Self-Regard",
                            "behaviour": "High Self-Regard Behavioural Insights",
                            "motivators": "High Self-Regard Motivators",
                            "strengths": "High Self-Regard Strengths",
                            "weaknesses": "High Self-Regard Weaknesses",
                            "reaction": "High Self-Regard Reaction Modes",
                            "communication": "High Self-Regard Communication"
                          },
                          {
                            "userID": 1,
                            "categoryName": "Self-Actualisation",
                            "score": 115,
                            "descName": "High Self-Actualisation",
                            "behaviour": "High Self-Actualisation Behavioural Insights",
                            "motivators": "High Self-Actualisation Motivators",
                            "strengths": "High Self-Actualisation Strengths",
                            "weaknesses": "High Self-Actualisation Weaknesses",
                            "reaction": "High Self-Actualisation Reaction Modes",
                            "communication": "High Self-Actualisation Communication"
                          },
                          {
                            "userID": 1,
                            "categoryName": "Emotional Self-Awareness",
                            "score": 115,
                            "descName": "High Emotional Self-Awareness",
                            "behaviour": "High Emotional Self-Awareness Behavioural Insights",
                            "motivators": "High Emotional Self-Awareness Motivators",
                            "strengths": "High Emotional Self-Awareness Strengths",
                            "weaknesses": "High Emotional Self-Awareness Weaknesses",
                            "reaction": "High Emotional Self-Awareness Reaction Modes",
                            "communication": "High Emotional Self-Awareness Communication"
                          },
                          {
                            "userID": 1,
                            "categoryName": "Emotional Expression",
                            "score": 115,
                            "descName": "High Emotional Expression",
                            "behaviour": "High Emotional Expression Behavioural Insights",
                            "motivators": "High Emotional Expression Motivators",
                            "strengths": "High Emotional Expression Strengths",
                            "weaknesses": "High Emotional Expression Weaknesses",
                            "reaction": "High Emotional Expression Reaction Modes",
                            "communication": "High Emotional Expression Communication"
                          },
                          {
                            "userID": 1,
                            "categoryName": "Asertiveness",
                            "score": 115,
                            "descName": "High Asertiveness",
                            "behaviour": "High Asertiveness Behavioural Insights",
                            "motivators": "High Asertiveness Motivators",
                            "strengths": "High Asertiveness Strengths",
                            "weaknesses": "High Asertiveness Weaknesses",
                            "reaction": "High Asertiveness Reaction Modes",
                            "communication": "High Asertiveness Communication"
                          },
                          {
                            "userID": 1,
                            "categoryName": "Independence",
                            "score": 115,
                            "descName": "High Independence",
                            "behaviour": "High Independence Behavioural Insights",
                            "motivators": "High Independence Motivators",
                            "strengths": "High Independence Strengths",
                            "weaknesses": "High Independence Weaknesses",
                            "reaction": "High Independence Reaction Modes",
                            "communication": "High Independence Communication"
                          },
                          {
                            "userID": 1,
                            "categoryName": "Interpersonal Relationships",
                            "score": 115,
                            "descName": "High Interpersonal Relationships",
                            "behaviour": "High Interpersonal Relationships Behavioural Insights",
                            "motivators": "High Interpersonal Relationships Motivators",
                            "strengths": "High Interpersonal Relationships Strengths",
                            "weaknesses": "High Interpersonal Relationships Weaknesses",
                            "reaction": "High Interpersonal Relationships Reaction Modes",
                            "communication": "High Interpersonal Relationships Communication"
                          },
                          {
                            "userID": 1,
                            "categoryName": "Empathy",
                            "score": 115,
                            "descName": "High Empathy",
                            "behaviour": "High Empathy Behavioural Insights",
                            "motivators": "High Empathy Motivators",
                            "strengths": "High Empathy Strengths",
                            "weaknesses": "High Empathy Weaknesses",
                            "reaction": "High Empathy Reaction Modes",
                            "communication": "High Empathy Communication"
                          },
                          {
                            "userID": 1,
                            "categoryName": "Social Responsibility",
                            "score": 115,
                            "descName": "High Social Responsibility",
                            "behaviour": "High Social Responsibility Behavioural Insights",
                            "motivators": "High Social Responsibility Motivators",
                            "strengths": "High Social Responsibility Strengths",
                            "weaknesses": "High Social Responsibility Weaknesses",
                            "reaction": "High Social Responsibility Reaction Modes",
                            "communication": "High Social Responsibility Communication"
                          },
                          {
                            "userID": 1,
                            "categoryName": "Problem Solving",
                            "score": 115,
                            "descName": "High Problem Solving",
                            "behaviour": "High Problem Solving Behavioural Insights",
                            "motivators": "High Problem Solving Motivators",
                            "strengths": "High Problem Solving Strengths",
                            "weaknesses": "High Problem Solving Weaknesses",
                            "reaction": "High Problem Solving Reaction Modes",
                            "communication": "High Problem Solving Communication"
                          },
                          {
                            "userID": 1,
                            "categoryName": "Reality Testing",
                            "score": 115,
                            "descName": "High Reality Testing",
                            "behaviour": "High Reality Testing Behavioural Insights",
                            "motivators": "High Reality Testing Motivators",
                            "strengths": "High Reality Testing Strengths",
                            "weaknesses": "High Reality Testing Weaknesses",
                            "reaction": "High Reality Testing Reaction Modes",
                            "communication": "High Reality Testing Communication"
                          },
                          {
                            "userID": 1,
                            "categoryName": "Impulse Control",
                            "score": 115,
                            "descName": "High Impulse Control",
                            "behaviour": "High Impulse Control Behavioural Insights",
                            "motivators": "High Impulse Control Motivators",
                            "strengths": "High Impulse Control Strengths",
                            "weaknesses": "High Impulse Control Weaknesses",
                            "reaction": "High Impulse Control Reaction Modes",
                            "communication": "High Impulse Control Communication"
                          },
                          {
                            "userID": 1,
                            "categoryName": "Flexibility",
                            "score": 115,
                            "descName": "High Flexibility",
                            "behaviour": "High Flexibility Behavioural Insights",
                            "motivators": "High Flexibility Motivators",
                            "strengths": "High Flexibility Strengths",
                            "weaknesses": "High Flexibility Weaknesses",
                            "reaction": "High Flexibility Reaction Modes",
                            "communication": "High Flexibility Communication"
                          },
                          {
                            "userID": 1,
                            "categoryName": "Stress Tolerance",
                            "score": 115,
                            "descName": "High Stress Tolerance",
                            "behaviour": "High Stress Tolerance Behavioural Insights",
                            "motivators": "High Stress Tolerance Motivators",
                            "strengths": "High Stress Tolerance Strengths",
                            "weaknesses": "High Stress Tolerance Weaknesses",
                            "reaction": "High Stress Tolerance Reaction Modes",
                            "communication": "High Stress Tolerance Communication"
                          },
                          {
                            "userID": 1,
                            "categoryName": "Optimism",
                            "score": 115,
                            "descName": "High Optimism",
                            "behaviour": "High Optimism Behavioural Insights",
                            "motivators": "High Optimism Motivators",
                            "strengths": "High Optimism Strengths",
                            "weaknesses": "High Optimism Weaknesses",
                            "reaction": "High Optimism Reaction Modes",
                            "communication": "High Optimism Communication"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "User results not found"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
}